+++ packages/lens-react/lib/components/Datatable/context/ContextInterfaces.ts
@@ -10,6 +10,12 @@ import {
 import { TableStylingInterface } from "../interface/TableStylingInterface";
 import { ActionsMappingPropInterface } from "../interface/ActionsMappingInterface";
 
+type CellRendererFn<T = any> = (rowData: T) => React.ReactNode;
+
+type CellRendererMap<T = any> = {
+  [columnKey: string]: CellRendererFn<T>;
+};
+
 interface CommonWrapperInterface {
   children: ReactNode;
 }
@@ -50,6 +56,7 @@ interface DatatableContextProviderInterface extends CommonWrapperInterface {
   disableReordering?: boolean;
   cellOverflow?: "wrap" | "clip" | "ellipsis";
   cellOverFlowStyles: string;
+  cellRenderer?: CellRendererMap;
 }
 
 interface DataTableLensContextProviderInterface
@@ -76,6 +83,7 @@ interface DatatableContextType
 }
 
 export type {
+  CellRendererMap,
   DatatableVariants,
   CommonWrapperInterface,
   DatatableContextType,